Add files via upload
Deep learning model trained on the S&P500 index. It use a Bernoulli Restricted Boltzmann machine and a classifier. Also, this model tell us to buy if the market goes down or to sell if the market goes up and an acurracy percentage.

Project initialy based from Alexandra Kropf's python for data science course.